+++ package.json
@@ -23,9 +23,11 @@
   },
   "dependencies": {
     "fs-extra": "^0.18.0",
+    "gulp": "^3.8.11",
     "gulp-inject": "^1.2.0",
     "gulp-util": "^3.0.4",
     "js-yaml": "^3.2.7",
+    "lodash.merge": "^3.2.1",
     "mkdirp": "^0.5.0",
     "node-sass": "^2.1.1",
     "q": "^1.2.0",
